Map generation part 2!
its really messy but it works. The map generation works in that we create a StringMap of chunks, all represented with a String. we use that to create something that looks kinda good and then convert it into a ChunkMap which is the StringMap but instead of Strings we have Chunks, and each one of these Chunks has its own internal map. We now combine all the internal maps into one big StringMap that will then get transfered into a TileMap which has yet to come. Wow! Thats a lot of work. Im glad I did it.